+++ user/ivi_features.c
@@ -25,11 +25,11 @@ void IviInit()
 
 void iv_config(uint8_t iv_test_mode, uint8_t iv_recovery_mode, uint8_t snb_state)
 {
-    Printf("%s IV test mode %s\r\n", iv_test_mode ? "Enable" : "Disable",
+    dprint("%s IV test mode %s\r\n", iv_test_mode ? "Enable" : "Disable",
     Cmd_mt_set_ivupdate_test_mode(iv_test_mode)->result ? "FAILED" : "SUCCESS");
-    Printf("Set IV Recovery Mode %s %s\r\n", iv_recovery_mode ? "Enable" : "Disable",
+    dprint("Set IV Recovery Mode %s %s\r\n", iv_recovery_mode ? "Enable" : "Disable",
     Cmd_mn_set_ivrecovery_mode(iv_recovery_mode)->result ? "FAILED" : "SUCCESS");
-    Printf("%s SNB %s\r\n", snb_state ? "Enable" : "Disable",
+    dprint("%s SNB %s\r\n", snb_state ? "Enable" : "Disable",
     Cmd_mt_set_local_config(mesh_node_beacon, 0, 1, &snb_state)->result ? "FAILED" : "SUCCESS");
 }
  
@@ -37,6 +37,7 @@ void iv_config(uint8_t iv_test_mode, uint8_t iv_recovery_mode, uint8_t snb_state
 //
 //
 //
+extern uint32 ShowSqueNum();
 void NodeIviUpdateProc(void)
 {
         pStageInfo = GetNodeStageInfo(NODE_IVI_UPDATE_PROC);
@@ -49,6 +50,13 @@ void NodeIviUpdateProc(void)
                 case IVI_DETECT: 
                     if(GetNodeStatus(NS_IVI_UPDATE) == ON)
                         {ToWaitingStage(IVI_UPDATE_ACTION,WAIT_SEC(1)); }
+                    else if(CheckWaitTimeOut()){
+                    	if (ShowSqueNum()<0x100){
+                    		ToWaitingStage(IVI_UPDATE_ACTION,WAIT_SEC(0));
+                    	}else{
+                    		ToWaitingStage(IVI_DETECT,WAIT_SEC(10));
+                    	}
+                    }
                     break;                   
                 case IVI_UPDATE_ACTION:  
                     IvIndexUpdate(ON);
@@ -85,35 +93,53 @@ Bool IvIndexUpdate(uchar status)
     Bool ret_code = FALSE;
     if(status == ON && GetMeshNodeStatus(STATUS_IVI_UPDATE) != ON)
     {
+    	dprint("!!!!! IvIndexUpdate(ON)\r\n");
         UsartIrq(USART_ID_RX,OFF);
         result = Cmd_mt_set_ivupdate_test_mode(ON)->result;
         if(result)
         {
+        	dprint("Cmd_mt_set_ivupdate_test_mode(ON):0x%x\r\n",result);
             return ret_code;
         }
+        /*dprint("!!! auto update iv index, current:%d (+%d)\r\n",pMeshNodeData->IvIndex,IVI_INC_MIN);
         result = Cmd_mt_set_iv_index(pMeshNodeData->IvIndex + IVI_INC_MIN)->result;
-        if(result){ return ret_code; }
+        if(result){
+        	dprint("Cmd_mt_set_iv_index(%d + %d):0x%x\r\n",pMeshNodeData->IvIndex,IVI_INC_MIN,result);
+        	return ret_code;
+        }
 
-        pMeshNodeData->IvIndex = pMeshNodeData->IvIndex + IVI_INC_MIN;
+        pMeshNodeData->IvIndex = pMeshNodeData->IvIndex + IVI_INC_MIN;*/
         result = Cmd_mn_req_ivupdate()->result;
-        if(result){return ret_code; }
+        if(result){
+        	dprint("Cmd_mn_req_ivupdate:0x%x\r\n",result);
+        	return ret_code;
+        }
         IviUpdateStatus(ON);
         result = Cmd_mt_send_beacons()->result;
-        if(result){return ret_code;}
+        if(result){
+        	dprint("Cmd_mt_send_beacons:0x%x\r\n",result);
+        	return ret_code;
+        }
         ret_code = TRUE;
     }
 
     if(status == OFF && GetMeshNodeStatus(STATUS_IVI_UPDATE) == ON)
     {
+    	dprint("!!!!! IvIndexUpdate(OFF)\r\n");
         result = Cmd_mt_set_ivupdate_state(OFF)->result;
-        if(result == bg_err_success)
-        {IviUpdateStatus(OFF);
-        result = Cmd_mt_send_beacons()->result;
-         if(result)
-         {return ret_code;}else 
-         {ret_code = TRUE;}
+        if(result == bg_err_success){
+        	IviUpdateStatus(OFF);
+        	result = Cmd_mt_send_beacons()->result;
+        	if(result){
+        		dprint("Cmd_mt_send_beacons:0x%x\r\n",result);
+        		return ret_code;
+        	}else{
+        		ret_code = TRUE;
+        	}
+        }else{
+        	dprint("Cmd_mt_set_ivupdate_state(OFF):0x%x\r\n",result);
         }
-      UsartIrq(USART_ID_RX,ON);
+        UsartIrq(USART_ID_RX,ON);
     }
     return ret_code;
 }
@@ -147,12 +173,21 @@ uint32 EvtMeshIviProc(PCmdPacket pEvent)
    
     uint32 ret_code = TRUE;
 
+    dprint("IV Recovery Needed !!!!\r\n");
+
     if(NodeRole == NR_CLIENT)
         EvtMeshIviClientProc(pEvent);
     else
         EvtMeshIviServerProc(pEvent);
 
     return ret_code;
+}
+
+uint32 ShowSqueNum(){
+    msg_mn_get_seq_remaining_rsp* seq_remain;
+	seq_remain=gecko_cmd_mesh_node_get_seq_remaining(0);
+	dprint("seq_remain, result:0x%x, count:%d(0x%X)\r\n",seq_remain->result,seq_remain->count,seq_remain->count);
+	return seq_remain->count;
 }
 
 
@@ -169,19 +204,28 @@ uint32 EvtMeshIviClientProc(PCmdPacket pEvent)
     switch(event_id)
     {
         case Evt_mn_ivrecovery_needed:
+        	dprint("!!!\r\n EVENT(Client) ivrecovery_needed\r\n>network IvIndex:%d\r\n node IvIndex:%d\r\n!!!\r\n",p_ivrecovery->network_ivindex,p_ivrecovery->node_ivindex);
             if(p_ivrecovery->network_ivindex != p_ivrecovery->node_ivindex)
-            { Cmd_mt_set_iv_index(p_ivrecovery->network_ivindex);
+            {
+            	result=Cmd_mt_set_iv_index(p_ivrecovery->network_ivindex)->result;
+            	dprint("set Iv Index: result:0x%X\r\n",result);
+            	if(!result) pMeshNodeData->IvIndex = p_ivrecovery->network_ivindex;
             }
             result = Cmd_mn_set_ivrecovery_mode(ON)->result;
+            ShowSqueNum();
             break;
 
-        case Evt_mn_changed_ivupdate_state:        
-            if(p_iv_update->state == ON)
-                {result = Cmd_mt_send_beacons()->result; 
-                }
-            else{result = Cmd_mt_send_beacons()->result;
-                }
+        case Evt_mn_changed_ivupdate_state:
+        	dprint("!!!\r\n EVENT(Client) change_ivupdate_state\r\n>state:%d,ivindex:%d\r\n!!!\r\n",p_iv_update->state,p_iv_update->ivindex);
+            if(p_iv_update->state == ON){
+            	result=Cmd_mt_set_iv_index(p_iv_update->ivindex)->result;
+            	dprint("set Iv Index: result:0x%X\r\n",result);
+            	result = Cmd_mt_send_beacons()->result;
+            }else{
+            	result = Cmd_mt_send_beacons()->result;
+            }
             pMeshNodeData->IvIndex = p_iv_update->ivindex;
+            ShowSqueNum();
             break;
 
     };
@@ -192,7 +236,8 @@ uint32 EvtMeshIviClientProc(PCmdPacket pEvent)
 
 uint32 EvtMeshIviServerProc(PCmdPacket pEvent)
 {
-    uint32 ret_code = TRUE;
+    //uint32 ret_code = TRUE;
+	uint16 result;
     uint32    event_id;
     msg_mn_ivrecovery_needed_evt *p_ivrecovery;
     msg_mn_changed_ivupdate_state_evt *p_iv_update;
@@ -203,30 +248,40 @@ uint32 EvtMeshIviServerProc(PCmdPacket pEvent)
     switch(event_id)
     {
         case Evt_mn_changed_ivupdate_state: 
+        	dprint("!!!\r\n EVENT(Server) change_ivupdate_state\r\n>state:%d,ivindex:%d\r\n!!!\r\n",p_iv_update->state,p_iv_update->ivindex);
             if(p_iv_update->state == YES){ 
-                IviUpdateStatus(ON);
+                //IviUpdateStatus(ON);
+            	result=Cmd_mt_set_iv_index(p_iv_update->ivindex)->result;
+            	if(!result) pMeshNodeData->IvIndex = p_iv_update->ivindex;
+            	dprint("set Iv Index: result:0x%X\r\n",result);
                 SetLedStatus(LED_STATUS_OFF);
             }
             else {
                 IviUpdateStatus(OFF);
                 SetLedStatus(LED_STATUS_IVI_UPDATE_OFF);
             }
+
+            ShowSqueNum();
             break;
 
         case Evt_mn_ivrecovery_needed: 
+        	dprint("!!!\r\n EVENT(Server) ivrecovery_needed\r\n>network IvIndex:%d\r\n node IvIndex:%d\r\n!!!\r\n",p_ivrecovery->network_ivindex,p_ivrecovery->node_ivindex);
             if(GetMeshNodeStatus(STATUS_PROXY_CONNECT) == TRUE)  break;
             if(p_ivrecovery->network_ivindex != p_ivrecovery->node_ivindex){                
-                Cmd_mt_set_iv_index(p_ivrecovery->network_ivindex);
+                result=Cmd_mt_set_iv_index(p_ivrecovery->network_ivindex)->result;
+                dprint("set Iv Index: result:0x%X\r\n",result);
                 IviUpdateStatus(OFF);
             }
             pMeshNodeData->IvIndex = p_ivrecovery->node_ivindex;
             result = Cmd_mn_set_ivrecovery_mode(ON)->result;
-            Cmd_mt_set_element_seqnum(PRIMARY_ELEM, 0); //set sequence num = 0
+            result=Cmd_mt_set_element_seqnum(PRIMARY_ELEM, 0)->result; //set sequence num = 0
+            dprint("Cmd_mt_set_element_seqnum:result:0x%X\r\n",result);
             SetMeshNodeStatus(STATUS_IVI_UPDATE, ON);
+            ShowSqueNum();
             break;
 
     };
-    return ret_code;
+    return TRUE;//ret_code;
 
 }